+++ src/bin/conserve.rs
@@ -61,6 +61,20 @@ enum Command {
         archive: PathBuf,
     },
 
+    /// Delete blocks unreferenced by any index.
+    ///
+    /// CAUTION: Do not gc while a backup is underway.
+    Gc {
+        /// Archive to delete from.
+        archive: PathBuf,
+        /// Don't actually delete, just check what could be deleted.
+        #[structopt(long)]
+        dry_run: bool,
+        /// Break a lock left behind by a previous interrupted gc operation, and then gc.
+        #[structopt(long)]
+        break_lock: bool,
+    },
+
     /// List files in a stored tree or source directory, with exclusions.
     Ls {
         #[structopt(flatten)]
@@ -205,6 +219,18 @@ impl Command {
                 let st = stored_tree_from_opt(archive, backup, exclude)?;
                 let lt = LiveTree::open(source)?.with_excludes(excludes);
                 output::show_tree_diff(&mut conserve::iter_merged_entries(&st, &lt)?, &mut stdout)?;
+            }
+            Command::Gc {
+                archive,
+                dry_run,
+                break_lock,
+            } => {
+                let archive = Archive::open_path(archive)?;
+                let stats = archive.delete_unreferenced(&DeleteOptions {
+                    dry_run: *dry_run,
+                    break_lock: *break_lock,
+                })?;
+                ui::println(&format!("{:#?}", stats));
             }
             Command::Init { archive } => {
                 Archive::create_path(&archive)?;